+++ src/Widget/TextArea.php
@@ -9,8 +9,8 @@ use Yiisoft\Widget\Widget;
 
 final class TextArea extends Widget
 {
-    use Collection\Options;
-    use Collection\InputOptions;
+    use Options\Common;
+    use Options\Input;
 
     /**
      * Generates a textarea tag for the given form attribute.
